requirements.txt
pure-python-adb==0.3.0.dev0
Pillow==9.1.0
pytesseract==0.3.9
